Fixes file transfer progress wrong data type (from int to long). Problem and fix thereof João Antunes!

+++ src/net/java/sip/communicator/impl/protocol/icq/FileTransferImpl.java
@@ -221,7 +221,7 @@ public class FileTransferImpl
 
                                 setTransferedBytes(transfered);
                                 fireProgressChangeEvent(
-                                    System.currentTimeMillis(), (int)transfered);
+                                    System.currentTimeMillis(), transfered);
 
                                 try {
                                     Thread.sleep(100);
